set cookies stored in `In::set_cookies`
users of this sdk can manipulate the `set_cookies`. This field will be
used by `ft_sdk::json_response` to set-cookie on the ongoing http
response

As a consequence of lifetimes in rust, a breaking change is added that
is, `impl`s of ft_sdk::Layout should not store reference to
`ft_sdk::In`. To come around this limitation and give them the data that
is stored in the `In` object, function params have been adjusted to
provide reference to `ft_sdk::In`. This is done for `Page::page()`,
`Action::validate()`, `Action::action()` and `Layout::from_in()`.