as a temp hack Query also looks in body
ideally it should be other way round, Required should look also in
query. But Query is simpler as it is implemented on String only,
whereas Required is generic.

Will implement proper generic solution for Required and then remove this
code, this code is only so we can make progress on our internal project.