remove canonical from browser_redirect_with_cookie
canonical is used to tell which URL is the right URL and should be used
in future, and is used when you change the url of a blog post. so this
equivalent to a permanent redirect. the browser_redirect_with_cookie,
at least in the use case we are using, is like temporary redirect, and
canonical is not appropriate there.